Fixing default message
