+++ src/main/java/com/example/Challenger2/entities/DTO/expenseDTOs/ExpenseDTO.java
@@ -1,7 +1,10 @@
-package com.example.Challenger2.entities.recipeDTOs;
+package com.example.Challenger2.entities.DTO.expenseDTOs;
 
-import com.example.Challenger2.entities.Recipe;
+import com.example.Challenger2.entities.Expense;
+import com.example.Challenger2.entities.enums.Category;
 import jakarta.persistence.Column;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -11,12 +14,14 @@ import lombok.*;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
+import static com.example.Challenger2.entities.enums.Category.getCategories;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode
-public class RecipeDTO {
+public class ExpenseDTO {
 
     @NotBlank(message = "Description is mandatory")
     private String description;
@@ -29,11 +34,16 @@ public class RecipeDTO {
     @NotNull(message = "Date is mandatory")
     @PastOrPresent(message = "Date must be in the past or present")
     private LocalDate date;
+    @NotNull(message = "Valid options are: Food, Health, Housing, Transport, Education, Leisure, Contingencies, Others")
+    @Enumerated(EnumType.STRING)
+    private Category category = Category.Others;
 
-    public RecipeDTO(Recipe recipe) {
-        this.description = recipe.getDescription();
-        this.price = recipe.getPrice();
-        this.date = recipe.getDate();
+    public ExpenseDTO(Expense expense) {
+        description = expense.getDescription();
+        price = expense.getPrice();
+        date = expense.getDate();
+        category = expense.getCategory();
     }
+
 
 }
