+++ src/main/java/com/example/Challenger2/entities/DTO/expenseDTOs/recipeDTOs/RecipeDTO.java
@@ -1,10 +1,7 @@
-package com.example.Challenger2.entities.expenseDTOs;
+package com.example.Challenger2.entities.DTO.expenseDTOs.recipeDTOs;
 
-import com.example.Challenger2.entities.Expense;
-import com.example.Challenger2.entities.enums.Category;
+import com.example.Challenger2.entities.Recipe;
 import jakarta.persistence.Column;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -14,14 +11,12 @@ import lombok.*;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-import static com.example.Challenger2.entities.enums.Category.getCategories;
-
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode
-public class ExpenseDTO {
+public class RecipeDTO {
 
     @NotBlank(message = "Description is mandatory")
     private String description;
@@ -34,16 +29,11 @@ public class ExpenseDTO {
     @NotNull(message = "Date is mandatory")
     @PastOrPresent(message = "Date must be in the past or present")
     private LocalDate date;
-    @NotNull()
-    @Enumerated(EnumType.STRING)
-    private Category category = Category.Outras;
 
-    public ExpenseDTO(Expense expense) {
-        description = expense.getDescription();
-        price = expense.getPrice();
-        date = expense.getDate();
-        category = expense.getCategory();
+    public RecipeDTO(Recipe recipe) {
+        this.description = recipe.getDescription();
+        this.price = recipe.getPrice();
+        this.date = recipe.getDate();
     }
-
 
 }
