+++ src/main/java/com/example/Challenger2/entities/Recipe.java
@@ -1,6 +1,6 @@
 package com.example.Challenger2.entities;
 
-import com.example.Challenger2.entities.recipeDTOs.RecipeDTO;
+import com.example.Challenger2.entities.DTO.expenseDTOs.recipeDTOs.RecipeDTO;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
