+++ src/main/java/com/example/Challenger2/services/ExpenseService.java
@@ -1,7 +1,7 @@
 package com.example.Challenger2.services;
 
 import com.example.Challenger2.entities.Expense;
-import com.example.Challenger2.entities.expenseDTOs.ExpenseDTO;
+import com.example.Challenger2.entities.DTO.expenseDTOs.ExpenseDTO;
 import com.example.Challenger2.repositories.ExpenseRepository;
 import com.example.Challenger2.services.exceptions.BadRequestException;
 import com.example.Challenger2.services.exceptions.NotFoundException;
