+++ src/main/resources/db/migration/V2__Update_expense_table.sql
@@ -1,2 +1,2 @@
 ALTER TABLE tb_expenses
-    ADD category VARCHAR(30);
+    ADD category VARCHAR(30) DEFAULT 'Others';
