--- src/main/resources/messages.properties
@@ -1,2 +1,0 @@
-javax.validation.constraints.NotNull.message=O campo {0} não pode ser nulo
-com.example.entities.Expense.category.NotNull={com.example.Challenger2.entities.enums.Category.getCategories};
